requirements.txt
qunetsim==0.1.2
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
networkx==3.1
cryptography==41.0.4
pytest==7.4.0
pyyaml==6.0.1
psutil==5.9.0  
